PreprocEmbedSPE: 首次 fp16-first 路徑重寫與 API 對齊
本次提交完成 PreprocEmbedSPE 區塊的主動路徑重寫，將計算流程切換為乾淨的 fp16-first 實作，並與 Top 層合約介面重新對齊。主要內容如下：

- 新增 PREPROC_REWRITE_ROUND2.md，詳細記錄重寫目的、影響、測試與現狀。
- 新增 weights.h 作為權重/header 包裝器，簡化 include 管理。
- 修正 fp16_from_double，提升 fp16 精度一致性。
- Fp16WeightProvider.h 改用新的權重 include 路徑。
- PreprocEmbedSPE.h 大幅重寫，保留原 API，內部改為 fp16-first 計算，恢復 Top 合約結構，新增 CoreWindow/Word16 介面，並保留 checkpoint copy 路徑供舊測試。
- 新增 gap probe 測試程式，協助比對 fp16 branch-side 與權威 RefModel 的 1 ulp 差異。

目前 Top 路徑已與 branch-side fp16 stage-local 完全一致，與 RefModel 尚有 1 ulp 差異，建議後續針對各階段插入比對點以釐清精度來源。

+++ include/Fp16RewriteTypes.h
@@ -50,7 +50,7 @@ static inline fp16_t fp16_from_word(const u16_t& word) {
 }
 
 static inline fp16_t fp16_from_double(const double value) {
-    return fp16_t((float)value);
+    return fp16_t(value);
 }
 
 static inline double fp16_to_double(const fp16_t& value) {
